creates new orgDetailViewController.swift and renames original to orgDetailViewControllerOld.swift
